+++ vescinterface.h
@@ -260,6 +260,8 @@ public:
         updateFwRx(false);
     }
 
+    Q_INVOKABLE bool downloadFwArchive();
+
 signals:
     void statusMessage(const QString &msg, bool isGood);
     void messageDialog(const QString &title, const QString &msg, bool isGood, bool richText);
@@ -279,6 +281,7 @@ signals:
     void configurationBackupsChanged();
     void customConfigLoadDone();
     void qmlLoadDone();
+    void fwArchiveDlProgress(QString msg, double prog);
 
 public slots: