+++ res/CHANGELOG.md
@@ -14,6 +14,7 @@
 * Automatically scan mobile CAN-list if empty.
 * Changelogs updated and changed to markdown.
 * Pixmap caching for faster start and config reload.
+* Firmware archive added in mobile VESC Tool.
 
 ---